refactor: reorganize tests into unit and e2e directories (#491)
# Reorganize Test Directory Structure for Better Separation of Unit and E2E Tests

This PR reorganizes the test directory structure to clearly separate unit tests from end-to-end tests:

- Moved all integration tests from `__tests__/integration/` to `__tests__/e2e/`
- Relocated unit tests from the root `__tests__/` directory to `__tests__/unit/`
- Updated import paths in test files to reflect the new directory structure
- Created a separate Vitest configuration for E2E tests (`vitest.e2e.config.ts`)
- Renamed `vitest.global-setup.ts` to `vitest.e2e.global-setup.ts` to clarify its purpose
- Updated the project.json configuration to use the appropriate test commands for each test type

This change improves the organization of our test suite and makes it clearer which tests are unit tests and which are end-to-end tests, while maintaining all existing test functionality.
