+++ pkgs/client/__tests__/e2e/full-stack-dsl.test.ts
