+++ pkgs/client/__tests__/e2e/happy-path-e2e.test.ts
