+++ pkgs/client/__tests__/e2e/network-resilience.test.ts
