+++ pkgs/client/__tests__/e2e/reconnection.test.ts
