+++ pkgs/client/__tests__/e2e/regressions/step-failed-event-bug.test.ts
