+++ pkgs/client/__tests__/e2e/supabase-setup.test.ts
