+++ pkgs/client/__tests__/e2e/wait-for-status-failure.test.ts
