+++ pkgs/client/vitest.e2e.config.ts
@@ -1,0 +1,12 @@
+/// <reference types='vitest' />
+import { defineConfig, mergeConfig } from 'vitest/config';
+import baseConfig from './vite.config';
+
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    test: {
+      globalSetup: './vitest.e2e.global-setup.ts',
+    },
+  })
+);
